LR1: moved from `Double` to `BigDecimal`

+++ LR1/app/src/main/java/com/example/unitconverter/converter/Converter.kt
@@ -1,6 +1,7 @@
 package com.example.unitconverter.converter
 
 import com.example.unitconverter.model.UnitItem
+import java.math.BigDecimal
 
 /**
  * Base interface for all converters
@@ -24,5 +25,5 @@ interface Converter {
      * @param toUnit the target unit
      * @return the converted value
      */
-    fun convert(value: Double, fromUnit: UnitItem, toUnit: UnitItem): Double
+    fun convert(value: BigDecimal, fromUnit: UnitItem, toUnit: UnitItem): BigDecimal
 }
